backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Database and Vector Storage
asyncpg>=0.29.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pgvector>=0.2.4
redis>=5.0.1

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.5.0
langgraph-checkpoint>=2.1.0
langgraph-prebuilt>=0.5.0
langgraph-sdk>=0.1.42
langsmith>=0.0.77

# Speech Recognition with Whisper
openai>=1.12.0
openai-whisper>=20231117

# Graphiti Knowledge Graph (Replaced with PostgreSQL RAG)
# graphiti-core>=0.3.0

# Document Processing
pypdf>=3.0.1
python-docx>=1.1.0
unstructured>=0.11.8
python-magic>=0.4.27

# Web Search and Scraping
tavily-python>=0.3.3
firecrawl-py>=2.13.0
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp>=3.9.1
websockets>=15.0.1
nest-asyncio>=1.6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.2
tenacity>=8.2.3
xxhash>=3.5.0
ormsgpack>=1.10.0
orjson>=3.10.1

# Security and Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring and Logging
structlog>=25.4.0
rich>=13.7.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1